DB::count(): improve compatibility
Addresses #10
The DB::count() function assumes that a query including `COUNT(*)` in the `SELECT` will return a column named `COUNT(*)`. This is not true in all database drivers! In Postgres, the column will be named `count`. In MSSQL, the column will be `unknown` (IIRC).
By fetching only 1 row (which is all that the query will return) using mode `PDO::FETCH_NUM`, we can avoid this issue completely.